move get_bad_snapshot to staticmethod in asg.py
